Problem: Business logic and representation logic are mixed togather in signer (CRO-72 / CRO-67) Solution: Extract all business logic of signer in a lib crate
