+++ Cargo.toml
@@ -3,6 +3,7 @@
 members = [
     "chain-abci",
     "chain-core",
-    "signer",
-    "dev-utils"
+    "dev-utils",
+    "signer-core",
+    "signer-cli",
 ]
