+++ signer-cli/src/commands.rs
