+++ signer-cli/src/signer.rs
