+++ signer-core/Cargo.toml
@@ -1,23 +1,16 @@
 [package]
-name = "signer"
+name = "signer-core"
 version = "0.1.0"
-authors = ["Crypto.com <chain@crypto.com>"]
-description = "Basic tool for secret management (address management, transaction generation) for Crypto.com chain node"
+authors = ["Devashish Dixit <devashish@crypto.com>"]
 edition = "2018"
 
 [dependencies]
 chain-core = { path = "../chain-core/" }
-structopt = "0.2"
-sled = "0.18"
-blake2 = "0.8"
-secp256k1zkp = { git = "https://github.com/crypto-com/rust-secp256k1-zkp.git", rev = "540cf449b16b4ee4874e9e211a50595e4141d67e", features = ["serde", "zeroize"] }
-miscreant = "^0.4.2"
-rand = "0.6"
-zeroize = "0.5"
-bincode = "1.1"
-serde = { version = "1.0", features = ["derive"] }
 sha3 = "0.8"
-hex = "0.3"
+secp256k1zkp = { git = "https://github.com/crypto-com/rust-secp256k1-zkp.git", rev = "540cf449b16b4ee4874e9e211a50595e4141d67e", features = ["serde", "zeroize"] }
+rand = "0.6"
 failure = "0.1"
-quest = "0.3"
-serde_cbor = "0.9"
+zeroize = "0.5"
+serde = { version = "1.0", features = ["derive"] }
+hex = "0.3"
+unicase = "2.2"
